worked on irg

+++ src/scanner.c
@@ -8,8 +8,8 @@
 
 int debug = 0;
 
-void lexing_error(size_t position, size_t line, char *code) {
-	printf("Error on char %lu in line %lu\n", position, line);
+void lexing_error(size_t position, long line, char *code) {
+	printf("Error on char %lu in line %ld\n", position, line);
 	printf("%.20s\n", code);
 	printf("^\n");
 	exit(1);
